SIG: move base to internal
