+++ go/sig/internal/base/pollhdlr.go
